Refactored and added score msg
