+++ src/Answer.jsx
@@ -2,7 +2,6 @@ import React from 'react'
 
 export default function Answer(props){
    // let isCorrectSubmitted = !props.isActive && props.isSelected && props.isCorrect
-   console.log(props)
 
     return (
         <p
