+++ src/App.js
@@ -1,22 +1,24 @@
 import './App.css';
 import React from 'react'
 import Start from './Start'
-import Quiz from './Quiz'
+import Question from './Question'
 import he from 'he'
+import {nanoid } from 'nanoid'
 import { ReactComponent as YellowBlob } from './blob-yellow.svg'
 import { ReactComponent as BlueBlob } from './blob-blue.svg'
 
 function App() {
+  
+  const [isActiveQuiz, setisActiveQuiz] = React.useState(true) //used to determine whether quiz is interactive or done/graded
+  const [questions, setQuestions] = React.useState([]) //used to store the array of questions from the API
+  const [quizId, setQuizId] = React.useState("") //used to generate new quizzes as appropriate in the useEffect dependency array
+  const [quizScore, setQuizScore] = React.useState(0) //used to keep track of how many questions are answered correctly
 
-  const [questions, setQuestions] = React.useState([])
-  const [isQuizStarted, setisQuizStarted] = React.useState(false)
-  const [quizCount, setQuizCount] = React.useState(1)
-
+  //Fetch 5 questions in any category of type: multiple choice from the Open Trivia DB API
   React.useEffect(()=>{
     fetch('https://opentdb.com/api.php?amount=5&category=9&type=multiple')
       .then(res => res.json())
       .then(data => {
-
         let questionsArray = data.results
         let questionObjects = questionsArray.map(question => {
           let answerArray = question.incorrect_answers.map(answerOption => {
@@ -32,26 +34,56 @@ function App() {
           let decodedQuestion = he.decode(question.question)
           return ({
             questionText: decodedQuestion,
-            answers: {answerArray}
+            answers: {answerArray},
+            isCorrect: false
           })
         })
         setQuestions(questionObjects)
       })
-  },[quizCount])
+  },[quizId])
 
   function startNewGame(){
-    setQuizCount(prevValue =>  {return (prevValue+1)})
-    setisQuizStarted(true)
+    setQuizId(nanoid())
+    setisActiveQuiz(true)
   }
 
+  //If the quiz is active, set it to inactive, but if it is already inactive, start a new quiz
+  function handleClick(){
+    isActiveQuiz ? setisActiveQuiz(false) : startNewGame()
+  }
+
+  function addCorrectToScore(questionText){
+      questions.find(question => question.questionText === questionText).isCorrect = true
+      setQuizScore(questions.filter(question => question.isCorrect === true).length)
+  }
+
+  //Get the array of Question components representing each question in this quiz
+  let questionDisplay = questions.map(question => {
+    return (<Question 
+                questionText={question.questionText} 
+                answers={question.answers} 
+                key={question.questionText} 
+                isActive={isActiveQuiz}
+                selectedAnswer={""}
+                setAsCorrect={addCorrectToScore}/>)
+  })
+
+//if there is not yet a quizId, display the start page, but if there is one, display the quiz
   return (
     <div className="App">
       <YellowBlob className="blob yellow"/>
       <BlueBlob className="blob blue"/>
       {
-        isQuizStarted 
+        quizId 
         ? 
-          <Quiz questions={questions} handleButtonClick={startNewGame}/>
+          <div className="quiz-container">
+              <h2>Quiz time</h2>
+              {questionDisplay}
+              <h4>{
+                !isActiveQuiz && `You scored ${quizScore}/5 correct answers` 
+              }</h4>
+              <button className='check-answers-btn' onClick={handleClick}>{isActiveQuiz ? 'Check answers' : 'Start new quiz'}</button>
+          </div>
         :
           <Start 
             handleClick={startNewGame}
